# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TrueNumber API",
    "version": "1.0.0",
    "description": "API documentation for the TrueNumber game application"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base API URL"
    }
  ],
  "paths": {
...
    "/user/{id}": {
      "get": {
        "summary": "Get user by ID (Admin only)",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User information"
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "put": {
        "summary": "Update user (Admin only)",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "role": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
    "user/history": {
      "get": {
        "summary": "Get user game history",
        "tags": ["History"],
        "responses": {
          "200": {
            "description": "Game history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "gameId": { "type": "string" },
                      "date": { "type": "string" },
                      "generatedNumber": { "type": "number" },
                      "result": { "type": "string" },
                      "balanceChange": { "type": "number" },
                      "newBalance": { "type": "number" }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      }
    }
  },
...
}